Split workload socket mapping into a module
